storage file include symbol aliases
The previous aliases no longer work as we are dealing with
objects which have symbol keys rather than hashes populated
with JSON data.

The previous aliases are kept around to protect backwards
compatability.

This seems to be the least awful method. Changing the aliases
is undesirable because that is another breaking change. Changing
the keys of each hash to be a string in order to match the
existing aliases is also unworkable as these can be deeply nested
and have names which are under different semantics(ie, camel vs
snake case)